fix: address workflow file action reviews
