+++ packages/extension/src/commands/files/openFileCommands.ts
@@ -40,7 +40,7 @@ export async function openFile(file: string, line?: number): Promise<void> {
   }
 }
 
-export async function openLabel(label: string): Promise<void> {
+export async function openLabel(label: string): Promise<boolean> {
   const escape = label.replaceAll(/[.*+?^${}()|[\]\\]/g, '\\$&');
   const pattern = new RegExp(`\\\\label\\{${escape}\\}`, 'm');
   const candidates = new Set([
@@ -60,7 +60,7 @@ export async function openLabel(label: string): Promise<void> {
           preview: true,
         });
         revealPosition(editor, doc.positionAt(match.index));
-        return;
+        return true;
       }
     } catch (error) {
       logger.debug(
@@ -71,6 +71,7 @@ export async function openLabel(label: string): Promise<void> {
   }
 
   vscode.window.showInformationMessage(`Label "${label}" not found.`);
+  return false;
 }
 
 export function registerOpenFileCommands(
